// src/types/formatSettings.ts
// Type for the format settings in the editor
export interface FormatSettings {
    spelling_and_grammar_check: boolean;
    paint_format: boolean;
    bold: boolean;
    italics: boolean;
    underline: boolean;
    font_size: number;
    font: string;
    line_height: number;
    vertical_margin: number;
    horizontal_margin: number;
}